change the time offset precision from int to long (#2448)
**Notes:**

The clockskew auto-correct logic in the SDK relies on the `int`
primitive type when calculating the offset. When the offset is converted
from milliseconds to days, the ms represented as an `int` have the
boundaries as -24 and +24 days. Changing it to long (64-bit precision)
fixes the limit.

+++ aws-android-sdk-core/src/main/java/com/amazonaws/AmazonWebServiceClient.java
@@ -95,7 +95,7 @@ public abstract class AmazonWebServiceClient {
 
     /** Optional offset (in seconds) to use when signing requests. */
     @SuppressWarnings("checkstyle:visibilitymodifier")
-    protected int timeOffset;
+    protected long timeOffset;
 
     /** AWS signer for authenticating requests. */
     private volatile Signer signer;
@@ -660,7 +660,7 @@ public abstract class AmazonWebServiceClient {
      *
      * @return The optional value for time offset (in seconds) for this client.
      */
-    public int getTimeOffset() {
+    public long getTimeOffset() {
         return timeOffset;
     }
